Add create comments when user is logged in
Created comments form when user is logged in only.
Added success message near the appropriate blog post and comment form.
Incremented the number of comments by one in the Comments Number display
Add helper in handlebars to determine if user in logged in and
deleted user in store when signing out.